# optimize get_queue query